+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -6,9 +6,11 @@
 
 ### Breaking Changes
 
+- `SearchIndexerSkillset`, `SearchField`, `SearchIndex`, `AnalyzeTextOptions`, `SearchResourceEncryptionKey`, `SynonymMap`, `SearchIndexerDataSourceConnection` are not longer subclasses of `_serialization.Model`.
+
 ### Bugs Fixed
 
-- Fixed the issue that `SearchIndexerSkillset`, `SearchField`, `SearchIndex`, etc. could not be serialized.
+- Fixed the issue that `SearchIndexerSkillset`, `SearchField`, `SearchIndex`, `AnalyzeTextOptions`, `SearchResourceEncryptionKey`, `SynonymMap`, `SearchIndexerDataSourceConnection` could not be serialized.
 
 ### Other Changes